Refractoring files
